+++ src/managednetworkfabric/azext_managednetworkfabric/aaz/latest/networkfabric/device/_resync_certificate.py
@@ -22,9 +22,9 @@ class ResyncCertificate(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2025-07-15",
+        "version": "2026-01-15-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.managednetworkfabric/networkdevices/{}/resynccertificates", "2025-07-15"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.managednetworkfabric/networkdevices/{}/resynccertificates", "2026-01-15-preview"],
         ]
     }
 
@@ -140,7 +140,7 @@ class ResyncCertificate(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-07-15",
+                    "api-version", "2026-01-15-preview",
                     required=True,
                 ),
             }